Create README.md
It is an emailer where we can send an email to our entire customer list with the daily weather and the schedule of the day. Gmail or any Mail server has an SMTP server that listens to request. The Project is developed for better understanding of Python and its application.